requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.26.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
fonttools==4.39.0
idna==3.4
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pywin32==305
pyzmq==25.0.0
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.6
